Fix int32/int64 mismatch in timed_message_ms

+++ src/editor_state_module.f90
@@ -145,7 +145,7 @@ module editor_state_module
 
         ! Timed status message (persists for ~2 seconds)
         character(len=256) :: timed_message = ''
-        integer(int32) :: timed_message_ms = 0  ! timestamp when set
+        integer(int64) :: timed_message_ms = 0  ! timestamp when set
     end type editor_state_t
 
 contains
